# Rewrite set_to_object() and related utility functions