@ivanka12 2018-06-23
обов'язково файл з тетстами  закидати?