@selurvedu 2018-06-27
Дуже бажано. Їх можна прив'язати до Travis CI і в майбутньому збільшувати покриття всього коду на Python в репозиторії.