@mykhaly 2018-06-29
Review status: 0 of 2 files reviewed, 4 unresolved discussions (waiting on @selurvedu and @ivanka12)

---

*[op_robot_tests/tests_files/service_keywords.py, line 466 at r3](https://reviewable.io/reviews/openprocurement/robot_tests/823#-LG9flJvANfcFa8EBbJa:-LG9flJw8z25Wc8LfgR_:b-mfryn4) ([raw file](https://github.com/openprocurement/robot_tests/blob/e8c49e4941a100bf05a78927b0980ae2bc9affcf/op_robot_tests/tests_files/service_keywords.py#L466)):*
> ```
> 'title must be one of %s' % str(STR_TYPES)
> ```

Це можна переписати через `format`. Як на мене, то форматування стрінги через `%` це занадто old school :)
І ерор я б починав з великої букви.

---

*[op_robot_tests/tests_files/service_keywords.py, line 476 at r3](https://reviewable.io/reviews/openprocurement/robot_tests/823#-LG9gU0e8g0GRuP80vJd:-LG9gU0e8g0GRuP80vJe:b-qwmrpf) ([raw file](https://github.com/openprocurement/robot_tests/blob/e8c49e4941a100bf05a78927b0980ae2bc9affcf/op_robot_tests/tests_files/service_keywords.py#L476)):*
<details><summary><i alt="summary">Quoted 7 lines of code…</i></summary>

> ```
>     description = obj.get('description', '')
>     if description:
>         if not isinstance(description, STR_TYPES):
>             raise TypeError('description must be one of %s' % str(STR_TYPES))
>         obj_id = re.match(regex, description)
>         if obj_id and len(obj_id.groups()) >= 1:
>             return obj_id.group(1)
> ```
</details>


Цей код - дублікація частини вище, за винятком того, що тут `description`, а там - `title`. Це може робити одна функція. Треба тільки придумати, як гарно перевірити return value тої функції.

---

*[op_robot_tests/tests_files/service_keywords.py, line 382 at r4](https://reviewable.io/reviews/openprocurement/robot_tests/823#-LG9h5wd94BxQvSZ4Odj:-LG9h5wd94BxQvSZ4Odk:bmgnye) ([raw file](https://github.com/openprocurement/robot_tests/blob/39cf34975d9c289c2ef4535c912e974716ac24b4/op_robot_tests/tests_files/service_keywords.py#L382)):*
> ```Python
> 
> 
> def set_to_object(obj, path, value):
> ```

Було б класно додати докстрінгу до функції, бо коду досить багато і логіка не тривіальна.

---

*[op_robot_tests/tests_files/service_keywords.py, line 422 at r4](https://reviewable.io/reviews/openprocurement/robot_tests/823#-LG9hkf56zaJ-2s1ewbV:-LG9hkf56zaJ-2s1ewbW:b-vl4cpl) ([raw file](https://github.com/openprocurement/robot_tests/blob/39cf34975d9c289c2ef4535c912e974716ac24b4/op_robot_tests/tests_files/service_keywords.py#L422)):*
> ```
>                     extension = [None for _ in
>                                  range(abs(index) + plusone - len(obj[key]))]
> ```

```
[None] * (abs(index) + plusone - len(obj[key]))
```

---


*Comments from [Reviewable](https://reviewable.io/reviews/openprocurement/robot_tests/823)*
<!-- Sent from Reviewable.io -->