@selurvedu 2018-07-03
Review status: 0 of 2 files reviewed, 4 unresolved discussions (waiting on @selurvedu, @mykhaly, and @ivanka12)

---

*[op_robot_tests/tests_files/service_keywords.py, line 466 at r3](https://reviewable.io/reviews/openprocurement/robot_tests/823#-LG9flJvANfcFa8EBbJa:-LGQxwWEBrKmyrg8YTOD:bg2bn09) ([raw file](https://github.com/openprocurement/robot_tests/blob/e8c49e4941a100bf05a78927b0980ae2bc9affcf/op_robot_tests/tests_files/service_keywords.py#L466)):*
1. For simple cases like this one, when there are a few arguments with no formatting expressions, `%` is simply more convenient when typing.

It also has a little performance advantage:

```bash
$ cat fmt.py
```
```python
#!/usr/bin/python2
# https://tomatohater.com/2013/08/30/python-performance-string-formatting/
from timeit import timeit
STR_TYPES = (str, unicode)
s = str(STR_TYPES)


def test_modulo():
    'title must be one of %s' % str(STR_TYPES)


def test_format():
    'title must be one of {}'.format(str(STR_TYPES))


def test_modulo_s():
    'title must be one of %s' % s


def test_format_s():
    'title must be one of {}'.format(s)


print(timeit(stmt=test_modulo))
print(timeit(stmt=test_format))
print(timeit(stmt=test_modulo_s))
print(timeit(stmt=test_format_s))
```
```bash
$ python2 fmt.py
1.87169718742
2.07400012016
0.25587105751
0.383615016937
```

2. I prefer the style of standard error messages in CPython. They are written in a non-capitalised manner (the first word is lowercase unless it is an abbreviation or some kind of an identifier), for example:

```python
>>> 1 / 0
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
ZeroDivisionError: integer division or modulo by zero
>>> '' + 1
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
TypeError: cannot concatenate 'str' and 'int' objects
>>> asdf
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
NameError: name 'asdf' is not defined
>>> x(])
  File "<stdin>", line 1
    x(])
      ^
SyntaxError: invalid syntax
>>> x(')
  File "<stdin>", line 1
    x(')
       ^
SyntaxError: EOL while scanning string literal
>>> # etc.
```

---

*[op_robot_tests/tests_files/service_keywords.py, line 476 at r3](https://reviewable.io/reviews/openprocurement/robot_tests/823#-LG9gU0e8g0GRuP80vJd:-LGR1J4h7fddfsQ7xARP:b-rl3dcn) ([raw file](https://github.com/openprocurement/robot_tests/blob/e8c49e4941a100bf05a78927b0980ae2bc9affcf/op_robot_tests/tests_files/service_keywords.py#L476)):*
<details><summary><i>Previously, mykhaly (Yurii Mykhalchuk) wrote…</i></summary><blockquote>

> ```
>     description = obj.get('description', '')
>     if description:
>         if not isinstance(description, STR_TYPES):
>             raise TypeError('description must be one of %s' % str(STR_TYPES))
>         obj_id = re.match(regex, description)
>         if obj_id and len(obj_id.groups()) >= 1:
>             return obj_id.group(1)
> ```

Цей код - дублікація частини вище, за винятком того, що тут `description`, а там - `title`. Це може робити одна функція. Треба тільки придумати, як гарно перевірити return value тої функції.
</blockquote></details>

I tried rewriting it a few times, but every time it looked uglier and uglier. You can take a look at my best attempt here (branch `prozorro/set_to_object-gifo`):

https://github.com/selurvedu/robot_tests/commit/2ce170c9aa40547b4226d8335f6a1d24f1dfa9cf

I am not adding it it yet to this PR since:
1) I don't want to postpone the merge of this PR;
2) the rewritten function needs to be covered with unit tests (as well as other functions in this PR), and I was going to add them later in another PR (because of point 1);
3) the rewritten function lacks a docstring.

---

*[op_robot_tests/tests_files/service_keywords.py, line 382 at r4](https://reviewable.io/reviews/openprocurement/robot_tests/823#-LG9h5wd94BxQvSZ4Odj:-LGQvVk0B-YsmoKuPi8N:b-bv222o) ([raw file](https://github.com/openprocurement/robot_tests/blob/39cf34975d9c289c2ef4535c912e974716ac24b4/op_robot_tests/tests_files/service_keywords.py#L382)):*
<details><summary><i>Previously, mykhaly (Yurii Mykhalchuk) wrote…</i></summary><blockquote>

Було б класно додати докстрінгу до функції, бо коду досить багато і логіка не тривіальна.
</blockquote></details>

I am being lazy to describe that complex logic. Even if you try to explain how it works, the explanation would not be easier than the function itself. :)
It worth describing what it is generally used for. I will prepare another PR that will add more unit tests and rework a couple more functions.

---

*[op_robot_tests/tests_files/service_keywords.py, line 422 at r4](https://reviewable.io/reviews/openprocurement/robot_tests/823#-LG9hkf56zaJ-2s1ewbV:-LGQlI-dC1AGx_GUwogu:b-n0l5bd) ([raw file](https://github.com/openprocurement/robot_tests/blob/39cf34975d9c289c2ef4535c912e974716ac24b4/op_robot_tests/tests_files/service_keywords.py#L422)):*
<details><summary><i>Previously, mykhaly (Yurii Mykhalchuk) wrote…</i></summary><blockquote>

> ```
>                     extension = [None for _ in
>                                  range(abs(index) + plusone - len(obj[key]))]
> ```

```
[None] * (abs(index) + plusone - len(obj[key]))
```
</blockquote></details>

You are right. Here is my brief research.

```bash
$ cat prof.py
```
```python
#!/usr/bin/python2
import profile


def main_old():
    key = 'foo'
    obj = {key: [1]}
    plusone = 0
    index = 99999999
    return [None for _ in range(abs(index) + plusone - len(obj[key]))]

def main_new():
    key = 'foo'
    obj = {key: [1]}
    plusone = 0
    index = 99999999
    return [None] * (abs(index) + plusone - len(obj[key]))


profile.run('old = main_old()')
profile.run('new = main_new()')
assert old == new
```
```bash
$ python2 prof.py
```
```python
         7 function calls in 8.379 seconds

   Ordered by: standard name

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
        1    0.000    0.000    0.000    0.000 :0(abs)
        1    0.000    0.000    0.000    0.000 :0(len)
        1    2.844    2.844    2.844    2.844 :0(range)
        1    0.001    0.001    0.001    0.001 :0(setprofile)
        1    0.000    0.000    8.378    8.378 <string>:1(<module>)
        1    5.534    5.534    8.378    8.378 prof.py:5(main_old)
        1    0.000    0.000    8.379    8.379 profile:0(old = main_old())
        0    0.000             0.000          profile:0(profiler)


         6 function calls in 0.894 seconds

   Ordered by: standard name

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
        1    0.000    0.000    0.000    0.000 :0(abs)
        1    0.000    0.000    0.000    0.000 :0(len)
        1    0.000    0.000    0.000    0.000 :0(setprofile)
        1    0.000    0.000    0.894    0.894 <string>:1(<module>)
        1    0.894    0.894    0.894    0.894 prof.py:12(main_new)
        1    0.000    0.000    0.894    0.894 profile:0(new = main_new())
        0    0.000             0.000          profile:0(profiler)

```

In the following example it is also four times more memory-efficient:

```bash
$ cat mem.py
```
```python
from __future__ import division


import gc
import os
import psutil


def main_old():
    return [None for _ in range(99999999)]


def main_new():
    return [None] * 99999999


process = psutil.Process(os.getpid())

print('Before new: %f MiB' % (process.memory_info().rss / 1024 / 1024))
new = main_new()
print('After new: %f MiB' % (process.memory_info().rss / 1024 / 1024))
del new
gc.collect()
print('After del new + gc and before old: %f MiB'
      % (process.memory_info().rss / 1024 / 1024))
old = main_old()
print('After old: %f MiB' % (process.memory_info().rss / 1024 / 1024))
del old
gc.collect()
print('After del old: %f MiB' % (process.memory_info().rss / 1024 / 1024))
```
```bash
$ python2 mem.py
Before new: 12.410156 MiB
After new: 775.277344 MiB
After del new + gc and before old: 12.457031 MiB
After old: 3120.187500 MiB
After del old: 2357.246094 MiB
```

Notice how the garbage is being left behind in memory in case of the old function.

---


*Comments from [Reviewable](https://reviewable.io/reviews/openprocurement/robot_tests/823)*
<!-- Sent from Reviewable.io -->